riskgroup refactor
