+++ app/src/utils/riskGroupsObject.js
@@ -1,0 +1,9 @@
+const riskGroups = {
+    dc: 'Doença respiratória',
+    hiv: 'HIV',
+    diab: 'Diabetes',
+    hiperT: 'Hipertensão',
+    doenCardio: 'Doenças cardiovasculares',
+};
+
+export default riskGroups;
